fix(#84): adds down migration and rename pkg entity to uniqueEntityId
