+++ pkg/uniqueEntityId/id_test.go
@@ -1,4 +1,4 @@
-package entity
+package uniqueEntityId
 
 import (
 	"testing"
